@@ -2,7 +2,13 @@
 {
     public class FieldInfo
     {
-        public string Type { get; set; }
-        public object Value { get; set; }
+        public FieldInfo(string type, object value)
+        {
+            Type = type;
+            Value = value;
+        }
+
+        public string Type { get; }
+        public object Value { get; }
     }
 }
