@@ -16,21 +16,28 @@ namespace QA.WidgetPlatform.Api.Models
         public string NodeType { get; set; }
         public IDictionary<string, FieldInfo> Details { get; set; }
 
-        public SiteNodeDetails(UniversalAbstractItem item, IEnumerable<string> excludeFields = null)
+        public SiteNodeDetails(UniversalAbstractItem item, IEnumerable<string>? excludeFields = null)
         {
             Id = item.Id;
             Alias = item.Alias;
             NodeType = item.Type;
 
             Details = item.UntypedFields
-                .Where(kvp => kvp.Value != null) // думаю, косяк в UniversalAbstractItem, отсекать null-значения скорее всего надо там
-                .Where(kvp => !Constants.AbstractItemSystemFields.Any(ef => ef.Equals(kvp.Key, StringComparison.InvariantCultureIgnoreCase)))
-                .Where(kvp => excludeFields == null || !excludeFields.Any(ef => ef.Equals(kvp.Key, StringComparison.InvariantCultureIgnoreCase)))
-                .ToDictionary(kvp => kvp.Key, kvp => new FieldInfo
-                {
-                    Value = kvp.Value,
-                    Type = kvp.Value.GetType().Name // думаю, нужно использовать справочник возможных типов qp, .net типы тут временно
-                });
+                .Where(kvp =>
+                    kvp.Value !=
+                    null) // думаю, косяк в UniversalAbstractItem, отсекать null-значения скорее всего надо там
+                .Where(kvp =>
+                    !Constants.AbstractItemSystemFields.Any(ef =>
+                        ef.Equals(kvp.Key, StringComparison.InvariantCultureIgnoreCase)))
+                .Where(kvp =>
+                    excludeFields == null || !excludeFields.Any(ef =>
+                        ef.Equals(kvp.Key, StringComparison.InvariantCultureIgnoreCase)))
+                .ToDictionary(kvp => kvp.Key, kvp =>
+                    new FieldInfo(
+                        // думаю, нужно использовать справочник возможных типов qp, .net типы тут временно
+                        kvp.Value.GetType().Name,
+                        kvp.Value
+                    ));
         }
     }
 }
